Refactor MCP tools,  gateway tools and Command_timeout usage on MCP client
