+++ hexstrike_mcp.py
@@ -11,16 +11,14 @@ Framework: FastMCP integration for tool orchestration
 
 import sys
 import argparse
-import json
 import logging
 from typing import Dict, Any, Optional
 import requests
 import time
 from datetime import datetime
+
 import core.config_core as config_core
-
 from mcp.server.fastmcp import FastMCP
-
 from mcp_core.hexstrikecolors import HexStrikeColors
 
 # Backward compatibility alias
@@ -73,7 +71,7 @@ logger = logging.getLogger(__name__)
 
 # Default configuration
 DEFAULT_HEXSTRIKE_SERVER = "http://127.0.0.1:8888"  # Default HexStrike server URL
-DEFAULT_REQUEST_TIMEOUT = 300  # 5 minutes default timeout for API requests
+DEFAULT_REQUEST_TIMEOUT = config_core.get("COMMAND_TIMEOUT", 300)  # 5 minutes default timeout
 MAX_RETRIES = 3  # Maximum number of retries for connection attempts
 
 class HexStrikeClient:
@@ -213,58 +211,9 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     # GATEWAY TOOLS (always registered)
     # ============================================================================
 
-    @mcp.tool()
-    def classify_task(description: str) -> Dict[str, Any]:
-        """
-        Classify a security task and return recommended tools.
-        Call this FIRST before running security tools to discover which ones are relevant.
-
-        Args:
-            description: What you want to do (e.g., "scan for open ports", "test for SQL injection")
-
-        Returns:
-            Task category, recommended tools with parameters, and usage instructions
-        """
-        result = hexstrike_client.safe_post("api/intelligence/classify-task", {"description": description})
-        if result.get("success"):
-            result["usage"] = "Use run_tool with a tool name and params from the recommended list"
-        return result
-
-    @mcp.tool()
-    def run_tool(tool_name: str, params: str) -> Dict[str, Any]:
-        """
-        Execute any security tool by name with parameters.
-        Use classify_task first to discover available tools.
-
-        Args:
-            tool_name: Tool name from classify_task results (e.g., "nmap", "nuclei")
-            params: JSON string of parameters (e.g., '{"target": "10.0.0.1"}')
-
-        Returns:
-            Tool execution results
-        """
-        from tool_registry import get_tool
-        tool_def = get_tool(tool_name)
-        if not tool_def:
-            return {"error": f"Unknown tool: {tool_name}", "success": False}
-
-        try:
-            parsed_params = json.loads(params) if isinstance(params, str) else params
-        except json.JSONDecodeError as e:
-            return {"error": f"Invalid params JSON: {e}", "success": False}
-
-        # Validate required params
-        for pname, spec in tool_def["params"].items():
-            if spec.get("required") and pname not in parsed_params:
-                return {"error": f"Missing required param: {pname}", "success": False}
-
-        # Fill defaults for optional params
-        for k, v in tool_def.get("optional", {}).items():
-            if k not in parsed_params:
-                parsed_params[k] = v
-
-        endpoint = tool_def["endpoint"].lstrip("/")
-        return hexstrike_client.safe_post(endpoint, parsed_params)
+    # Register gateway tools for task classification and tool execution
+    from mcp_tools.gateway import register_gateway_tools
+    register_gateway_tools(mcp, hexstrike_client)
 
     if compact:
         logger.info("Compact mode: only gateway tools registered (classify_task, run_tool)")
